refactor(week): refining reminder code to Reminder.ts
