+++ .prettierrc.yml
@@ -1,4 +1,4 @@
 trailingComma: "es5"
-tabWidth: 4
 useTabs: true
+tabWidth: 4
 semi: true
