+++ src/configuration.ts
@@ -1,5 +1,3 @@
-"use strict";
-
 import * as vscode from "vscode";
 
 const ReminderAheadTime = "bangumiOpen.ReminderAheadTime";
