+++ src/test/runTest.ts
@@ -14,7 +14,10 @@ async function main() {
 		const extensionTestsPath = path.resolve(__dirname, "./suite/index");
 
 		// Download VS Code, unzip it and run the integration test
-		await runTests({ extensionDevelopmentPath, extensionTestsPath });
+		await runTests({
+			extensionDevelopmentPath,
+			extensionTestsPath,
+		});
 	} catch (err) {
 		console.error("Failed to run tests");
 		process.exit(1);
