+++ src/utils/file.ts
@@ -1,5 +1,3 @@
-"use strict";
-
 import * as vscode from "vscode";
 import * as path from "path";
 import { getContext } from "../constants";
