+++ src/utils/type.ts
@@ -1,5 +1,3 @@
-"use strict";
-
 /**
  * Detroit: Become Number
  *
